requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

numpy=1.19.2=py37h54aff64_0
pandas=1.2.0=py37ha9443f7_0
pip=20.3.3=py37h06a4308_0
python=3.7.9=h7579374_0
pytorch=1.7.1=py3.7_cuda10.2.89_cudnn7.6.5_0
pyyaml=5.3.1=py37h7b6447c_1
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.5.2=py37h0b6359f_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu102